+++ NEWS.md
@@ -45,7 +45,6 @@ hux(head(iris)) %>% set_bold(1, everywhere)
 * New: by default, huxtable sets labels from the current knitr chunk label, if there
   is one. This is consistent with `kable()`. In bookdown, you can then do e.g.
 
-
 ````
 Some iris species are shown in \@ref(tab:mytable):
 
@@ -56,8 +55,11 @@ hux(head(iris))
 
   Set `options(huxtable.autolabel = FALSE)` to return to the old behaviour.
   
+* The one-argument form of `[` now works for huxtables just as it does for 
+  data frames. For example, `ht[2:3]` selects columns 2 and 3.
 * New functions `merge_across()` and `merge_down()` to create multicolumn
   or multirow cells.
+* huxtable supports `dplyr::relocate()`, new in dplyr 1.0.0.
 * huxtable no longer supports dplyr versions less than 0.7.0 (released in mid-2017).
 * Improvements to `as_flextable()`.
 * Improvements to `quick_pptx()` (thanks @davidgohel).